Refactor SHOW SHARDING ALGORITHM IMPLEMENTATIONS
