+++ docs/document/content/user-manual/shardingsphere-proxy/distsql/syntax/ral/show-implementation/sharding/show-sharding-algorithm-implementations.cn.md
@@ -23,11 +23,11 @@ showShardingAlgorithmImplementations ::=
 
 ### 返回值说明
 
-| 列    | 说明      |
-|------|---------|
-| name | 实现类名称   |
-| type | 类型      |
-| class_path | 实现类完整路径 |
+| 列            | 说明     |
+|--------------|--------|
+| type         | 类型     |
+| type_aliases | 类型别名   |
+| description  | 描述     |
 
 ### 示例
 
@@ -39,20 +39,20 @@ SHOW SHARDING ALGORITHM IMPLEMENTATIONS
 
 ```sql
 SHOW SHARDING ALGORITHM IMPLEMENTATIONS;
-+-------------------------------------+----------------+-------------------------------------------------------------------------------------------------+
-| name                                | type           | class_path                                                                                      |
-+-------------------------------------+----------------+-------------------------------------------------------------------------------------------------+
-| ModShardingAlgorithm                | MOD            | org.apache.shardingsphere.sharding.algorithm.sharding.mod.ModShardingAlgorithm                  |
-| HashModShardingAlgorithm            | HASH_MOD       | org.apache.shardingsphere.sharding.algorithm.sharding.mod.HashModShardingAlgorithm              |
-| VolumeBasedRangeShardingAlgorithm   | VOLUME_RANGE   | org.apache.shardingsphere.sharding.algorithm.sharding.range.VolumeBasedRangeShardingAlgorithm   |
-| BoundaryBasedRangeShardingAlgorithm | BOUNDARY_RANGE | org.apache.shardingsphere.sharding.algorithm.sharding.range.BoundaryBasedRangeShardingAlgorithm |
-| AutoIntervalShardingAlgorithm       | AUTO_INTERVAL  | org.apache.shardingsphere.sharding.algorithm.sharding.datetime.AutoIntervalShardingAlgorithm    |
-| IntervalShardingAlgorithm           | INTERVAL       | org.apache.shardingsphere.sharding.algorithm.sharding.datetime.IntervalShardingAlgorithm        |
-| ClassBasedShardingAlgorithm         | CLASS_BASED    | org.apache.shardingsphere.sharding.algorithm.sharding.classbased.ClassBasedShardingAlgorithm    |
-| InlineShardingAlgorithm             | INLINE         | org.apache.shardingsphere.sharding.algorithm.sharding.inline.InlineShardingAlgorithm            |
-| ComplexInlineShardingAlgorithm      | COMPLEX_INLINE | org.apache.shardingsphere.sharding.algorithm.sharding.inline.ComplexInlineShardingAlgorithm     |
-| HintInlineShardingAlgorithm         | HINT_INLINE    | org.apache.shardingsphere.sharding.algorithm.sharding.hint.HintInlineShardingAlgorithm          |
-+-------------------------------------+----------------+-------------------------------------------------------------------------------------------------+
++----------------+--------------+-------------+
+| type           | type_aliases | description |
++----------------+--------------+-------------+
+| MOD            |              |             |
+| HASH_MOD       |              |             |
+| VOLUME_RANGE   |              |             |
+| BOUNDARY_RANGE |              |             |
+| AUTO_INTERVAL  |              |             |
+| INTERVAL       |              |             |
+| CLASS_BASED    |              |             |
+| INLINE         |              |             |
+| COMPLEX_INLINE |              |             |
+| HINT_INLINE    |              |             |
++----------------+--------------+-------------+
 10 rows in set (0.27 sec)
 ```
