+++ features/sharding/distsql/handler/src/test/java/org/apache/shardingsphere/sharding/distsql/query/ShowShardingAlgorithmImplementationsResultRowBuilderTest.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.sharding.distsql.query;
 
 import org.apache.shardingsphere.infra.merge.result.impl.local.LocalDataQueryResultRow;
-import org.apache.shardingsphere.sharding.distsql.handler.query.ShardingSPIImplementationsResultRowBuilder;
+import org.apache.shardingsphere.sharding.distsql.handler.query.ShowShardingAlgorithmImplementationsResultRowBuilder;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collection;
@@ -27,11 +27,11 @@ import java.util.Iterator;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-class ShardingSPIImplementationsResultRowBuilderTest {
+class ShowShardingAlgorithmImplementationsResultRowBuilderTest {
     
     @Test
     void assertGetRowData() {
-        ShardingSPIImplementationsResultRowBuilder rowBuilder = new ShardingSPIImplementationsResultRowBuilder();
+        ShowShardingAlgorithmImplementationsResultRowBuilder rowBuilder = new ShowShardingAlgorithmImplementationsResultRowBuilder();
         Collection<LocalDataQueryResultRow> actual = rowBuilder.generateRows();
         assertThat(actual.size(), is(15));
         Iterator<LocalDataQueryResultRow> iterator = actual.iterator();
